TI_AI_SaaS_Project/requirements.txt
Django==5.2.9
celery==5.4.0
redis==7.1.0
selenium==4.38.0
webdriver-manager==4.0.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt>=5.5.1
djoser>=2.3.0
social-auth-app-django>=5.6.0
python-dotenv==1.0.0
django-celery-beat==2.8.1
django-cors-headers==4.2.0
django-csp==3.7
langchain>=1.1.0,<2.0.0
langgraph>=1.0.2,<2.0.0
pypdf==5.1.0
python-docx==1.1.2
shadcn-django==0.23.0
argon2-cffi==23.1.0
channels==4.3.2
channels-redis==4.3.0
uuid6==2025.0.1
nanoid==2.0.0
django-storages==1.14.0
boto3==1.34.0
phonenumbers==8.13.0
email-validator==2.1.0
django-extensions==3.2.3
graphviz==0.20.3
pydot==2.0.0